+++ app/upgrades/v19/upgrades.go
@@ -30,6 +30,8 @@ const (
 	OsmosisOutpostAddress = "0x0000000000000000000000000000000000000901"
 )
 
+var newExtraEIPs = []string{"evmos_0", "evmos_1", "evmos_2"}
+
 // CreateUpgradeHandler creates an SDK upgrade handler for v19
 func CreateUpgradeHandler(
 	mm *module.Manager,
@@ -46,11 +48,6 @@ func CreateUpgradeHandler(
 		logger.Debug("deleting revenue module from version map...")
 		delete(vm, "revenue")
 
-		ctxCache, writeFn := ctx.CacheContext()
-		if err := RemoveOutpostsFromEvmParams(ctxCache, ek); err == nil {
-			writeFn()
-		}
-
 		MigrateEthAccountsToBaseAccounts(ctx, ak, ek)
 
 		// run module migrations first.
@@ -58,6 +55,13 @@ func CreateUpgradeHandler(
 		migrationRes, err := mm.RunMigrations(ctx, configurator, vm)
 		if err != nil {
 			return migrationRes, err
+		}
+
+		ctxCache, writeFn := ctx.CacheContext()
+		if err := RemoveOutpostsFromEvmParams(ctxCache, ek); err == nil {
+			writeFn()
+		} else {
+			logger.Error("error removing outposts", "error", err)
 		}
 
 		bondDenom := sk.BondDenom(ctx)
@@ -77,6 +81,12 @@ func CreateUpgradeHandler(
 			writeFn()
 		}
 
+		ctxCache, writeFn = ctx.CacheContext()
+		if err := EnableCustomEIPs(ctxCache, logger, ek); err == nil {
+			writeFn()
+		} else {
+			logger.Error("error setting new extra EIPs", "error", err)
+		}
 		return migrationRes, err
 	}
 }
@@ -203,4 +213,20 @@ func MigrateEthAccountsToBaseAccounts(ctx sdk.Context, ak authkeeper.AccountKeep
 
 		return false
 	})
+}
+
+func EnableCustomEIPs(ctx sdk.Context, logger log.Logger, ek *evmkeeper.Keeper) error {
+	params := ek.GetParams(ctx)
+	extraEIPs := params.ExtraEIPs
+
+	for _, eip := range newExtraEIPs {
+		if slices.Contains(extraEIPs, eip) {
+			logger.Debug("skipping duplicate EIP", "eip", eip)
+		} else {
+			extraEIPs = append(extraEIPs, eip)
+		}
+	}
+
+	params.ExtraEIPs = extraEIPs
+	return ek.SetParams(ctx, params)
 }