+++ x/evm/migrations/v7/migrate.go
@@ -1,18 +1,22 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
 package v7
 
 import (
+	"fmt"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/evmos/evmos/v19/x/evm/types"
 
 	v6types "github.com/evmos/evmos/v19/x/evm/migrations/v7/types"
+	"github.com/evmos/evmos/v19/x/evm/types"
 )
 
 // MigrateStore migrates the x/evm module state from the consensus version 6 to
-// version 7. Specifically, it adds the new AccessControl policy.
+// version 7. Specifically, it changes the type of the Params ExtraEIPs from
+// []int64 to []string and introduces the access control.
 func MigrateStore(
 	ctx sdk.Context,
 	storeKey storetypes.StoreKey,
@@ -29,7 +33,6 @@ func MigrateStore(
 	cdc.MustUnmarshal(paramsV6Bz, &paramsV6)
 
 	params.EvmDenom = paramsV6.EvmDenom
-	params.ExtraEIPs = paramsV6.ExtraEIPs
 	params.ChainConfig = types.ChainConfig{
 		HomesteadBlock:      paramsV6.ChainConfig.HomesteadBlock,
 		DAOForkBlock:        paramsV6.ChainConfig.DAOForkBlock,
@@ -58,6 +61,15 @@ func MigrateStore(
 	// set the default access control configuration
 	params.AccessControl = types.DefaultAccessControl
 
+	// Migrate old ExtraEIPs from int64 to string. Since no Evmos EIPs have been
+	// created before and activators contains only `ethereum_XXXX` activations,
+	// all values will be prefixed with `ethereum_`.
+	params.ExtraEIPs = make([]string, 0, len(paramsV6.ExtraEIPs))
+	for _, eip := range paramsV6.ExtraEIPs {
+		eipName := fmt.Sprintf("ethereum_%d", eip)
+		params.ExtraEIPs = append(params.ExtraEIPs, eipName)
+	}
+
 	if err := params.Validate(); err != nil {
 		return err
 	}
@@ -65,5 +77,6 @@ func MigrateStore(
 	bz := cdc.MustMarshal(&params)
 
 	store.Set(types.KeyPrefixParams, bz)
+
 	return nil
 }